Split test_models into separate files
